Delete the file containing the app
An updated version of the app will be uploaded. 

This is done via the deletion because the update was produced on a local PC and it contains too many new features. At this point, it's more efficient to delete the current version and add an upgraded one. In future, code editing will be processed via GitHub commits.