Add napp.py file
This file contains the updated version of the app, including some bug fixes.

New features added:
- An interval with weekly prices has been added. 
- RSI indicator has been modified so that any value of RSI can be chosen. (e.g. RSI-7, RSI-14)
- Splits are now processed for SBER, SBERP, MGNT, NVTK tickers. 
- Tickers can now be input in lowercase.
- "Enter"/"Return" key can now be used to enter any inputs.
- The GUI has been modified to accommodate for the new changes. 

Bug Fixes:
- The SMA is now calculated by 60 rows. (Before is was 1260 days, less accurate)

-----
The update is presented by deletion and a consequent upload of the app.py file. This is because the update was created locally and includes many features. At this point, It's more efficient to upload it this way. 

Future changes will be processed via GitHub commits for convenience.